Move datastore package
