+++ package.json
@@ -11,6 +11,7 @@
     "lodash": "~4.17.4",
     "loglevel": "~1.5.1",
     "qs": "~6.5.1",
+    "request": "~2.83.0",
     "rxjs": "~5.5.2",
     "sift": "~5.0.0"
   },
@@ -18,8 +19,10 @@
     "babel-preset-env": "~1.6.1",
     "babel-register": "~6.26.0",
     "chai": "~4.1.2",
+    "chai-as-promised": "~7.1.1",
     "expect": "~1.20.2",
     "local-storage": "~1.4.2",
-    "mocha": "~4.0.1"
+    "mocha": "~4.0.1",
+    "proxyquire": "~1.8.0"
   }
 }
