--- packages/kinvey-datastore/package.json
@@ -1,32 +1,0 @@
-{
-  "private": true,
-  "version": "3.9.3-beta.7",
-  "name": "kinvey-datastore",
-  "main": "src/index.js",
-  "dependencies": {
-    "es6-promise": "4.1.1",
-    "kinvey-aggregation": "^3.9.3-beta.7",
-    "kinvey-errors": "^3.9.3-beta.7",
-    "kinvey-live": "^3.9.3-beta.7",
-    "kinvey-metadata": "^3.9.3-beta.7",
-    "kinvey-observable": "^3.9.3-beta.7",
-    "kinvey-query": "^3.9.3-beta.7",
-    "kinvey-request": "^3.9.3-beta.7",
-    "kinvey-utils": "^3.9.3-beta.7",
-    "lodash": "4.17.4"
-  },
-  "devDependencies": {
-    "chai": "4.1.2",
-    "chai-as-promised": "7.1.1",
-    "expect": "^1.20.2",
-    "kinvey": "^3.9.3-beta.7",
-    "kinvey-user": "^3.9.3-beta.7",
-    "mocha": "4.0.1",
-    "nock": "9.0.28",
-    "proxyquire": "1.8.0",
-    "request": "2.83.0"
-  },
-  "scripts": {
-    "test": "mocha -s 250 test/**/*.spec.js"
-  }
-}
