--- packages/kinvey-datastore/src/index.js
@@ -1,14 +1,0 @@
-const { CacheStore } = require('./cachestore');
-const { NetworkStore } = require('./networkstore');
-const { SyncStore } = require('./syncstore');
-const { DataStoreType, DataStore } = require('./datastore');
-const { SyncOperation } = require('./sync');
-
-module.exports = {
-  SyncOperation: SyncOperation,
-  CacheStore: CacheStore,
-  NetworkStore: NetworkStore,
-  SyncStore: SyncStore,
-  DataStoreType: DataStoreType,
-  DataStore: DataStore
-};
