+++ src/datastore/index.js
@@ -1,0 +1,5 @@
+export * from './cachestore';
+export * from './networkstore';
+export * from './syncstore';
+export * from './datastore';
+export * from './sync';
