+++ src/datastore/syncstore.spec.js
@@ -1,22 +1,21 @@
-const { CacheStore, SyncStore, SyncOperation } = require('../src');
-const { init } = require('kinvey');
-const { Query } = require('kinvey-query');
-const { Aggregation } = require('kinvey-aggregation');
-const { KinveyError, NotFoundError, ServerError } = require('kinvey-errors');
-const { randomString } = require('kinvey-utils/string');
-const { NetworkRack } = require('kinvey-request');
-const { User } = require('kinvey-user');
-const { HttpMiddleware } = require('./http');
-const nock = require('nock');
-const expect = require('expect');
+import nock from 'nock';
+import expect from 'expect';
+import { SyncStore, SyncOperation } from './sync';
+import { Aggregation } from '../aggregation';
+import { Query } from '../query';
+import { KinveyError, NotFoundError } from '../errors';
+import { randomString } from '../utils';
+import { NetworkRack, NodeHttpMiddleware } from '../request';
+import { User } from '../user';
+import { init } from '../kinvey';
 
 const collection = 'Books';
 
-describe('CacheStore', () => {
+describe('SyncStore', () => {
   let client;
 
   before(() => {
-    NetworkRack.useHttpMiddleware(new HttpMiddleware());
+    NetworkRack.useHttpMiddleware(new NodeHttpMiddleware({}));
   });
 
   before(() => {
@@ -50,7 +49,7 @@ describe('CacheStore', () => {
   });
 
   afterEach(() => {
-    const store = new CacheStore(collection);
+    const store = new SyncStore(collection);
     return store.clear()
       .then(() => {
         return store.clearSync();
@@ -59,12 +58,12 @@ describe('CacheStore', () => {
 
   describe('pathname', () => {
     it(`should equal /appdata/<appkey>/${collection}`, () => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       expect(store.pathname).toEqual(`/appdata/${client.appKey}/${collection}`);
     });
 
     it('should not be able to be changed', () => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       store.pathname = `/tests/${collection}`;
       expect(store.pathname).toEqual(`/appdata/${client.appKey}/${collection}`);
     });
@@ -72,20 +71,20 @@ describe('CacheStore', () => {
 
   describe('syncAutomatically', () => {
     it('should be true', () => {
-      const store = new CacheStore(collection);
-      expect(store.syncAutomatically).toEqual(true);
+      const store = new SyncStore(collection);
+      expect(store.syncAutomatically).toEqual(false);
     });
 
     it('should not be able to be changed', () => {
-      const store = new CacheStore(collection);
-      store.syncAutomatically = false;
-      expect(store.syncAutomatically).toEqual(true);
+      const store = new SyncStore(collection);
+      store.syncAutomatically = true;
+      expect(store.syncAutomatically).toEqual(false);
     });
   });
 
   describe('find()', () => {
     it('should throw an error if the query argument is not an instance of the Query class', (done) => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       store.find({})
         .subscribe(null, (error) => {
           try {
@@ -100,81 +99,35 @@ describe('CacheStore', () => {
         });
     });
 
-    it('should throw an error if there are entities to sync', (done) => {
-      const entity = { _id: randomString() };
-      const syncStore = new SyncStore(collection);
-      syncStore.save(entity)
-        .then(() => {
-          const store = new CacheStore(collection);
-          store.find()
-            .subscribe(null, (error) => {
-              try {
-                expect(error).toBeA(KinveyError);
-                expect(error.message).toEqual(
-                  'Unable to fetch the entities on the backend.'
-                  + ' There are 1 entities that need to be synced.'
-                );
-                done();
-              } catch (e) {
-                done(e);
-              }
-            }, () => {
-              done(new Error('This test should fail.'));
-            });
-        });
-    });
-
     it('should return the entities', (done) => {
       const entity1 = { _id: randomString() };
       const entity2 = { _id: randomString() };
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const onNextSpy = expect.createSpy();
 
       nock(store.client.apiHostname)
-        .get(store.pathname)
+        .get(`/appdata/${store.client.appKey}/${collection}`)
         .reply(200, [entity1, entity2]);
 
       store.pull()
         .then(() => {
-          const entity3 = {
-            _id: randomString()
-          };
-
-          nock(store.client.apiHostname)
-            .get(store.pathname)
-            .reply(200, [entity1, entity2, entity3]);
-
           store.find()
             .subscribe(onNextSpy, done, () => {
               try {
-                expect(onNextSpy.calls.length).toEqual(2);
+                expect(onNextSpy.calls.length).toEqual(1);
                 expect(onNextSpy.calls[0].arguments).toEqual([[entity1, entity2]]);
-                expect(onNextSpy.calls[1].arguments).toEqual([[entity1, entity2, entity3]]);
-
-                onNextSpy.reset();
-                const syncStore = new SyncStore(collection);
-                syncStore.find()
-                  .subscribe(onNextSpy, done, () => {
-                    try {
-                      expect(onNextSpy.calls.length).toEqual(1);
-                      expect(onNextSpy.calls[0].arguments).toEqual([[entity1, entity2, entity3]]);
-                      done();
-                    } catch (error) {
-                      done(error);
-                    }
-                  });
+                done();
               } catch (error) {
                 done(error);
               }
             });
-        })
-        .catch(done);
+        });
     });
 
     it('should return the entities that match the query', (done) => {
       const entity1 = { _id: randomString() };
       const entity2 = { _id: randomString() };
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const query = new Query().equalTo('_id', entity1._id);
       const onNextSpy = expect.createSpy();
 
@@ -184,64 +137,12 @@ describe('CacheStore', () => {
 
       store.pull()
         .then(() => {
-          nock(store.client.apiHostname)
-            .get(`/appdata/${store.client.appKey}/${collection}`)
-            .query({ query: JSON.stringify({ _id: entity1._id }) })
-            .reply(200, [entity1]);
-
           store.find(query)
             .subscribe(onNextSpy, done, () => {
               try {
-                expect(onNextSpy.calls.length).toEqual(2);
+                expect(onNextSpy.calls.length).toEqual(1);
                 expect(onNextSpy.calls[0].arguments).toEqual([[entity1]]);
-                expect(onNextSpy.calls[1].arguments).toEqual([[entity1]]);
                 done();
-              } catch (error) {
-                done(error);
-              }
-            });
-        });
-    });
-
-    it('should remove entities that no longer exist on the backend from the cache', (done) => {
-      const entity1 = { _id: randomString() };
-      const entity2 = { _id: randomString() };
-      const store = new CacheStore(collection);
-      const onNextSpy = expect.createSpy();
-
-      nock(store.client.apiHostname)
-        .get(`/appdata/${store.client.appKey}/${collection}`)
-        .reply(200, [entity1, entity2]);
-
-      store.pull()
-        .then(() => {
-          const entity3 = {
-            _id: randomString()
-          };
-
-          nock(store.client.apiHostname)
-            .get(`/appdata/${store.client.appKey}/${collection}`)
-            .reply(200, [entity1, entity3]);
-
-          store.find()
-            .subscribe(onNextSpy, done, () => {
-              try {
-                expect(onNextSpy.calls.length).toEqual(2);
-                expect(onNextSpy.calls[0].arguments).toEqual([[entity1, entity2]]);
-                expect(onNextSpy.calls[1].arguments).toEqual([[entity1, entity3]]);
-
-                onNextSpy.reset();
-                const syncStore = new SyncStore(collection);
-                syncStore.find()
-                  .subscribe(onNextSpy, done, () => {
-                    try {
-                      expect(onNextSpy.calls.length).toEqual(1);
-                      expect(onNextSpy.calls[0].arguments).toEqual([[entity1, entity3]]);
-                      done();
-                    } catch (error) {
-                      done(error);
-                    }
-                  });
               } catch (error) {
                 done(error);
               }
@@ -252,7 +153,7 @@ describe('CacheStore', () => {
 
   describe('findById()', () => {
     it('should return undefined if an id is not provided', (done) => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const onNextSpy = expect.createSpy();
       store.findById()
         .subscribe(onNextSpy, done, () => {
@@ -266,44 +167,15 @@ describe('CacheStore', () => {
         });
     });
 
-    it('should throw an error if there are entities to sync', (done) => {
-      const entity = { _id: randomString() };
-      const syncStore = new SyncStore(collection);
-      syncStore.save(entity)
-        .then(() => {
-          const store = new CacheStore(collection);
-          store.findById(entity._id)
-            .subscribe(null, (error) => {
-              try {
-                expect(error).toBeA(KinveyError);
-                expect(error.message).toEqual(
-                  'Unable to find the entity on the backend.'
-                  + ' There are 1 entities that need to be synced.'
-                );
-                done();
-              } catch (e) {
-                done(e);
-              }
-            }, () => {
-              done(new Error('This test should fail.'));
-            });
-        });
-    });
-
     it('should throw a NotFoundError if the entity does not exist', (done) => {
       const entity = { _id: randomString() };
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const onNextSpy = expect.createSpy();
-
-      nock(store.client.apiHostname)
-        .get(`/appdata/${store.client.appKey}/${collection}/${entity._id}`)
-        .reply(404);
 
       store.findById(entity._id)
         .subscribe(onNextSpy, (error) => {
           try {
-            expect(onNextSpy.calls.length).toEqual(1);
-            expect(onNextSpy.calls[0].arguments).toEqual([undefined]);
+            expect(onNextSpy.calls.length).toEqual(0);
             expect(error).toBeA(NotFoundError);
             done();
           } catch (e) {
@@ -317,7 +189,7 @@ describe('CacheStore', () => {
     it('should return the entity that matches the id', (done) => {
       const entity1 = { _id: randomString() };
       const entity2 = { _id: randomString() };
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const onNextSpy = expect.createSpy();
 
       nock(store.client.apiHostname)
@@ -326,16 +198,11 @@ describe('CacheStore', () => {
 
       store.pull()
         .then(() => {
-          nock(store.client.apiHostname)
-            .get(`/appdata/${store.client.appKey}/${collection}/${entity1._id}`)
-            .reply(200, entity1);
-
           store.findById(entity1._id)
             .subscribe(onNextSpy, done, () => {
               try {
-                expect(onNextSpy.calls.length).toEqual(2);
+                expect(onNextSpy.calls.length).toEqual(1);
                 expect(onNextSpy.calls[0].arguments).toEqual([entity1]);
-                expect(onNextSpy.calls[1].arguments).toEqual([entity1]);
                 done();
               } catch (error) {
                 done(error);
@@ -343,43 +210,11 @@ describe('CacheStore', () => {
             });
         });
     });
-
-    it('should remove entities that no longer exist on the backend from the cache', (done) => {
-      const entity1 = { _id: randomString() };
-      const entity2 = { _id: randomString() };
-      const store = new CacheStore(collection);
-      const onNextSpy = expect.createSpy();
-
-      nock(store.client.apiHostname)
-        .get(`/appdata/${store.client.appKey}/${collection}`)
-        .reply(200, [entity1, entity2]);
-
-      store.pull()
-        .then(() => {
-          nock(store.client.apiHostname)
-            .get(`/appdata/${store.client.appKey}/${collection}/${entity1._id}`)
-            .reply(404);
-
-          store.findById(entity1._id)
-            .subscribe(onNextSpy, (error) => {
-              try {
-                expect(onNextSpy.calls.length).toEqual(1);
-                expect(onNextSpy.calls[0].arguments).toEqual([entity1]);
-                expect(error).toBeA(NotFoundError);
-                done();
-              } catch (e) {
-                done(e);
-              }
-            }, () => {
-              done(new Error('This test should fail.'));
-            });
-        });
-    });
   });
 
   describe('group()', () => {
     it('should throw an error if the query argument is not an instance of the Query class', (done) => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       store.group({})
         .subscribe(null, (error) => {
           try {
@@ -394,56 +229,10 @@ describe('CacheStore', () => {
         });
     });
 
-    it('should throw a ServerError', (done) => {
-      const store = new CacheStore(collection);
-      const aggregation = new Aggregation();
-
-      nock(client.apiHostname)
-        .post(`/appdata/${client.appKey}/${collection}/_group`)
-        .reply(500);
-
-      store.group(aggregation)
-        .subscribe(null, (error) => {
-          try {
-            expect(error).toBeA(ServerError);
-            done();
-          } catch (e) {
-            done(e);
-          }
-        }, () => {
-          done(new Error('This test should fail.'));
-        });
-    });
-
-    it('should throw an error if there are entities to sync', (done) => {
-      const entity = { _id: randomString() };
-      const syncStore = new SyncStore(collection);
-      syncStore.save(entity)
-        .then(() => {
-          const aggregation = new Aggregation();
-          const store = new CacheStore(collection);
-          store.group(aggregation)
-            .subscribe(null, (error) => {
-              try {
-                expect(error).toBeA(KinveyError);
-                expect(error.message).toEqual(
-                  'Unable to group entities on the backend.'
-                  + ' There are 1 entities that need to be synced.'
-                );
-                done();
-              } catch (e) {
-                done(e);
-              }
-            }, () => {
-              done(new Error('This test should fail.'));
-            });
-        });
-    });
-
     it('should return the count of all unique properties on the collection', (done) => {
       const entity1 = { _id: randomString(), title: randomString() };
       const entity2 = { _id: randomString(), title: randomString() };
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const onNextSpy = expect.createSpy();
 
       nock(store.client.apiHostname)
@@ -453,17 +242,11 @@ describe('CacheStore', () => {
       store.pull()
         .then(() => {
           const aggregation = Aggregation.count('title');
-          const reply = [{ title: randomString(), count: 2 }, { title: randomString(), count: 1 }];
-          nock(client.apiHostname)
-            .post(`/appdata/${client.appKey}/${collection}/_group`)
-            .reply(200, reply);
-
           store.group(aggregation)
             .subscribe(onNextSpy, done, () => {
               try {
-                expect(onNextSpy.calls.length).toEqual(2);
+                expect(onNextSpy.calls.length).toEqual(1);
                 expect(onNextSpy.calls[0].arguments).toEqual([[{ count: 1, title: entity1.title }, { count: 1, title: entity2.title }]]);
-                expect(onNextSpy.calls[1].arguments).toEqual([reply]);
                 done();
               } catch (error) {
                 done(error);
@@ -475,7 +258,7 @@ describe('CacheStore', () => {
 
   describe('count()', () => {
     it('should throw an error if the query argument is not an instance of the Query class', (done) => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       store.count({})
         .subscribe(null, (error) => {
           try {
@@ -490,54 +273,10 @@ describe('CacheStore', () => {
         });
     });
 
-    it('should throw a ServerError', (done) => {
-      const store = new CacheStore(collection);
-
-      nock(client.apiHostname)
-        .get(`/appdata/${client.appKey}/${collection}/_count`)
-        .reply(500);
-
-      store.count()
-        .subscribe(null, (error) => {
-          try {
-            expect(error).toBeA(ServerError);
-            done();
-          } catch (e) {
-            done(e);
-          }
-        }, () => {
-          done(new Error('This test should fail.'));
-        });
-    });
-
-    it('should throw an error if there are entities to sync', (done) => {
-      const entity = { _id: randomString() };
-      const syncStore = new SyncStore(collection);
-      syncStore.save(entity)
-        .then(() => {
-          const store = new CacheStore(collection);
-          store.count()
-            .subscribe(null, (error) => {
-              try {
-                expect(error).toBeA(KinveyError);
-                expect(error.message).toEqual(
-                  'Unable to count entities on the backend.'
-                  + ' There are 1 entities that need to be synced.'
-                );
-                done();
-              } catch (e) {
-                done(e);
-              }
-            }, () => {
-              done(new Error('This test should fail.'));
-            });
-        });
-    });
-
     it('should return the count for the collection', (done) => {
       const entity1 = { _id: randomString() };
       const entity2 = { _id: randomString() };
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const onNextSpy = expect.createSpy();
 
       nock(store.client.apiHostname)
@@ -546,16 +285,11 @@ describe('CacheStore', () => {
 
       store.pull()
         .then(() => {
-          nock(client.apiHostname)
-            .get(`/appdata/${client.appKey}/${collection}/_count`)
-            .reply(200, { count: 3 });
-
           store.count()
             .subscribe(onNextSpy, done, () => {
               try {
-                expect(onNextSpy.calls.length).toEqual(2);
+                expect(onNextSpy.calls.length).toEqual(1);
                 expect(onNextSpy.calls[0].arguments).toEqual([2]);
-                expect(onNextSpy.calls[1].arguments).toEqual([3]);
                 done();
               } catch (error) {
                 done(error);
@@ -567,7 +301,7 @@ describe('CacheStore', () => {
 
   describe('create()', () => {
     it('should throw an error if trying to create an array of entities', () => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const entity1 = {};
       const entity2 = {};
 
@@ -579,60 +313,46 @@ describe('CacheStore', () => {
     });
 
     it('should create an entity', () => {
-      const store = new CacheStore(collection);
-      const entity = { title: randomString() };
-      const reply = { _id: randomString(), title: entity.title };
-
-      nock(client.apiHostname)
-        .post(`/appdata/${client.appKey}/${collection}`, (body) => {
-          return body.title === entity.title;
-        })
-        .reply(201, reply);
-
+      const store = new SyncStore(collection);
+      const entity = {};
       return store.create(entity)
         .then((createdEntity) => {
-          expect(createdEntity).toEqual(reply);
+          expect(createdEntity).toIncludeKey('_id');
+          expect(createdEntity.title).toEqual(entity.title);
 
           // Check the cache to make sure the entity was
           // stored in the cache
-          const syncStore = new SyncStore(collection);
           const query = new Query();
           query.equalTo('_id', createdEntity._id);
-          return syncStore.find(query).toPromise()
+          return store.find(query).toPromise()
             .then((entities) => {
               expect(entities).toEqual([createdEntity]);
               return store.pendingSyncCount();
             })
             .then((count) => {
-              expect(count).toEqual(0);
+              expect(count).toEqual(1);
             });
         });
     });
 
     it('should create an entity if it contains an _id', async () => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const entity = { _id: randomString() };
-
-      nock(client.apiHostname)
-        .post(`/appdata/${client.appKey}/${collection}`, entity)
-        .reply(200, entity);
-
       return store.create(entity)
         .then((createdEntity) => {
           expect(createdEntity).toEqual(entity);
 
           // Check the cache to make sure the entity was
           // stored in the cache
-          const syncStore = new SyncStore(collection);
           const query = new Query();
           query.equalTo('_id', createdEntity._id);
-          return syncStore.find(query).toPromise()
+          return store.find(query).toPromise()
             .then((entities) => {
               expect(entities).toEqual([createdEntity]);
               return store.pendingSyncCount();
             })
             .then((count) => {
-              expect(count).toEqual(0);
+              expect(count).toEqual(1);
             });
         });
     });
@@ -640,7 +360,7 @@ describe('CacheStore', () => {
 
   describe('update()', () => {
     it('should throw an error if trying to update an array of entities', async () => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const entity1 = { _id: randomString() };
       const entity2 = { _id: randomString() };
 
@@ -652,7 +372,7 @@ describe('CacheStore', () => {
     });
 
     it('should throw an error if an entity does not have an _id', async () => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const entity = {};
 
       return store.update(entity)
@@ -664,29 +384,23 @@ describe('CacheStore', () => {
     });
 
     it('should update an entity with an _id', async () => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const entity = { _id: randomString() };
-
-      nock(client.apiHostname)
-        .put(`/appdata/${client.appKey}/${collection}/${entity._id}`, entity)
-        .reply(200, entity);
-
       return store.update(entity)
         .then((updatedEntity) => {
           expect(updatedEntity).toEqual(entity);
 
           // Check the cache to make sure the entity was
           // stored in the cache
-          const syncStore = new SyncStore(collection);
           const query = new Query();
           query.equalTo('_id', updatedEntity._id);
-          return syncStore.find(query).toPromise()
+          return store.find(query).toPromise()
             .then((entities) => {
               expect(entities).toEqual([updatedEntity]);
               return store.pendingSyncCount();
             })
             .then((count) => {
-              expect(count).toEqual(0);
+              expect(count).toEqual(1);
             });
         });
     });
@@ -698,21 +412,30 @@ describe('CacheStore', () => {
     });
 
     it('should call create() for an entity that does not contain an _id', () => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const spy = expect.spyOn(store, 'create');
       store.save({});
       expect(spy).toHaveBeenCalled();
     });
 
     it('should call update() for an entity that contains an _id', () => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const spy = expect.spyOn(store, 'update');
       store.save({ _id: randomString() });
       expect(spy).toHaveBeenCalled();
     });
 
+    it('should call update() for an entity that contains an _id with special characters', () => {
+      const store = new SyncStore(collection);
+      const id = '.$~<>!@+_#';
+      return store.save({ _id: id })
+        .then((resp) => {
+          expect(resp._id).toEqual(id);
+        });
+    });
+
     it('should call create() when an array of entities is provided', () => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const spy = expect.spyOn(store, 'create');
       store.save([{ _id: randomString() }, {}]);
       expect(spy).toHaveBeenCalled();
@@ -721,7 +444,7 @@ describe('CacheStore', () => {
 
   describe('remove()', () => {
     it('should throw an error if the query argument is not an instance of the Query class', () => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       store.remove({})
         .catch((error) => {
           expect(error).toBeA(KinveyError);
@@ -733,7 +456,7 @@ describe('CacheStore', () => {
     });
 
     it('should return a { count: 0 } when no entities are removed', () => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       return store.remove()
         .then((result) => {
           expect(result).toEqual({ count: 0 });
@@ -741,7 +464,7 @@ describe('CacheStore', () => {
     });
 
     it('should remove all the entities', () => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const entity1 = { _id: randomString() };
       const entity2 = { _id: randomString() };
 
@@ -751,32 +474,23 @@ describe('CacheStore', () => {
 
       return store.pull()
         .then(() => {
-          nock(store.client.apiHostname)
-            .delete(`/appdata/${store.client.appKey}/${collection}/${entity1._id}`)
-            .reply(200);
-
-          nock(store.client.apiHostname)
-            .delete(`/appdata/${store.client.appKey}/${collection}/${entity2._id}`)
-            .reply(200);
-
           return store.remove();
         })
         .then((result) => {
           expect(result).toEqual({ count: 2 });
-          const syncStore = new SyncStore(collection);
-          return syncStore.find().toPromise();
+          return store.find().toPromise();
         })
         .then((entities) => {
           expect(entities).toEqual([]);
           return store.pendingSyncCount();
         })
         .then((count) => {
-          expect(count).toEqual(0);
+          expect(count).toEqual(2);
         });
     });
 
     it('should remove all the entities that match the query', () => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const entity1 = { _id: randomString() };
       const entity2 = { _id: randomString() };
 
@@ -787,101 +501,25 @@ describe('CacheStore', () => {
       return store.pull()
         .then(() => {
           const query = new Query().equalTo('_id', entity1._id);
-
-          nock(store.client.apiHostname)
-            .delete(`/appdata/${store.client.appKey}/${collection}/${entity1._id}`)
-            .reply(200);
-
           return store.remove(query);
         })
         .then((result) => {
           expect(result).toEqual({ count: 1 });
-          const syncStore = new SyncStore(collection);
-          return syncStore.find().toPromise();
+          return store.find().toPromise();
         })
         .then((entities) => {
           expect(entities).toEqual([entity2]);
           return store.pendingSyncCount();
         })
         .then((count) => {
-          expect(count).toEqual(0);
-        });
-    });
-
-    it('should not remove the entity from the cache if the backend request failed', () => {
-      const store = new CacheStore(collection);
-      const entity1 = { _id: randomString() };
-      const entity2 = { _id: randomString() };
-
-      nock(store.client.apiHostname)
-        .get(`/appdata/${store.client.appKey}/${collection}`)
-        .reply(200, [entity1, entity2]);
-
-      return store.pull()
-        .then(() => {
-          const query = new Query().equalTo('_id', entity1._id);
-
-          nock(store.client.apiHostname)
-            .delete(`/appdata/${store.client.appKey}/${collection}/${entity1._id}`)
-            .reply(500);
-
-          return store.remove(query);
-        })
-        .then((result) => {
-          expect(result).toEqual({ count: 0 });
-          const syncStore = new SyncStore(collection);
-          return syncStore.find().toPromise();
-        })
-        .then((entities) => {
-          expect(entities).toEqual([entity1, entity2]);
-          return store.pendingSyncCount();
-        })
-        .then((count) => {
           expect(count).toEqual(1);
-        });
-    });
-
-    it('should remove an entity if it was created locally', () => {
-      const store = new CacheStore(collection);
-      const syncStore = new SyncStore(collection);
-      const entity1 = { _id: randomString() };
-      const entity2 = { _id: randomString() };
-      const entity3 = {};
-
-      nock(store.client.apiHostname)
-        .get(`/appdata/${store.client.appKey}/${collection}`)
-        .reply(200, [entity1, entity2]);
-
-      return store.pull()
-        .then(() => syncStore.save(entity3))
-        .then(() => {
-          nock(store.client.apiHostname)
-            .delete(`${store.pathname}/${entity1._id}`)
-            .reply(200);
-
-          nock(store.client.apiHostname)
-            .delete(`${store.pathname}/${entity2._id}`)
-            .reply(200);
-
-          return store.remove();
-        })
-        .then((result) => {
-          expect(result).toEqual({ count: 3 });
-          return syncStore.find().toPromise();
-        })
-        .then((entities) => {
-          expect(entities).toEqual([]);
-          return store.pendingSyncCount();
-        })
-        .then((count) => {
-          expect(count).toEqual(0);
         });
     });
   });
 
   describe('removeById()', () => {
-    it('should return a { count: 0 } if an id is not provided', () => {
-      const store = new CacheStore(collection);
+    it('should return a `{ count: 0 } if an id is not provided', () => {
+      const store = new SyncStore(collection);
       return store.removeById()
         .then((result) => {
           expect(result).toEqual({ count: 0 });
@@ -889,7 +527,7 @@ describe('CacheStore', () => {
     });
 
     it('should remove the entity from cache if the entity is not found on the backend', () => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const entity = { _id: randomString() };
 
       nock(store.client.apiHostname)
@@ -898,44 +536,12 @@ describe('CacheStore', () => {
 
       return store.pull()
         .then(() => {
-          nock(store.client.apiHostname)
-            .delete(`/appdata/${store.client.appKey}/${collection}/${entity._id}`)
-            .reply(404);
-
           return store.removeById(entity._id);
         })
         .then((result) => {
           expect(result).toEqual({ count: 1 });
-          const syncStore = new SyncStore(collection);
           const query = new Query().equalTo('_id', entity._id);
-          return syncStore.find(query).toPromise();
-        })
-        .then((entities) => {
-          expect(entities).toEqual([]);
-        });
-    });
-
-    it('should remove the entity from the backend', () => {
-      const store = new CacheStore(collection);
-      const entity = { _id: randomString() };
-
-      nock(store.client.apiHostname)
-        .get(`/appdata/${store.client.appKey}/${collection}`)
-        .reply(200, [entity]);
-
-      return store.pull()
-        .then(() => {
-          nock(store.client.apiHostname)
-            .delete(`/appdata/${store.client.appKey}/${collection}/${entity._id}`)
-            .reply(200, { count: 1 });
-
-          return store.removeById(entity._id);
-        })
-        .then((result) => {
-          expect(result).toEqual({ count: 1 });
-          const syncStore = new SyncStore(collection);
-          const query = new Query().equalTo('_id', entity._id);
-          return syncStore.find(query).toPromise();
+          return store.find(query).toPromise();
         })
         .then((entities) => {
           expect(entities).toEqual([]);
@@ -943,16 +549,17 @@ describe('CacheStore', () => {
     });
 
     it('should remove the entity from the cache and sync table', () => {
-      const store = new CacheStore(collection);
-      const syncStore = new SyncStore(collection);
+      const store = new SyncStore(collection);
       const entity = {};
 
-      return syncStore.save(entity)
-        .then(entity => store.removeById(entity._id))
+      return store.save(entity)
+        .then((entity) => {
+          return store.removeById(entity._id);
+        })
         .then((result) => {
           expect(result).toEqual({ count: 1 });
           const query = new Query().equalTo('_id', entity._id);
-          return syncStore.find(query).toPromise();
+          return store.find(query).toPromise();
         })
         .then((entities) => {
           expect(entities).toEqual([]);
@@ -967,7 +574,7 @@ describe('CacheStore', () => {
 
   describe('clear()', () => {
     it('should remove all entities only from the cache', () => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const entity = { _id: randomString() };
 
       nock(store.client.apiHostname)
@@ -980,8 +587,7 @@ describe('CacheStore', () => {
         })
         .then((result) => {
           expect(result).toEqual({ count: 1 });
-          const syncStore = new SyncStore(collection);
-          return syncStore.find().toPromise();
+          return store.find().toPromise();
         })
         .then((entities) => {
           expect(entities).toEqual([]);
@@ -989,7 +595,7 @@ describe('CacheStore', () => {
     });
 
     it('should remove only the entities from the cache that match the query', () => {
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
       const entity1 = { _id: randomString() };
       const entity2 = { _id: randomString() };
 
@@ -1004,23 +610,52 @@ describe('CacheStore', () => {
         })
         .then((result) => {
           expect(result).toEqual({ count: 1 });
-          const syncStore = new SyncStore(collection);
           const query = new Query().equalTo('_id', entity1._id);
-          return syncStore.find(query).toPromise();
+          return store.find(query).toPromise();
         })
         .then((entities) => {
           expect(entities).toEqual([]);
+        });
+    });
+
+    it('should remove an entity if it was created locally and not add it to the sync queue', () => {
+      const store = new SyncStore(collection);
+      const entity1 = { _id: randomString() };
+      const entity2 = { _id: randomString() };
+      const entity3 = {};
+
+      nock(store.client.apiHostname)
+        .get(`/appdata/${store.client.appKey}/${collection}`)
+        .reply(200, [entity1, entity2]);
+
+      return store.pull()
+        .then(() => {
+          return store.save(entity3);
+        })
+        .then(() => {
+          return store.remove();
+        })
+        .then((result) => {
+          expect(result).toEqual({ count: 3 });
+          const syncStore = new SyncStore(collection);
+          return syncStore.find().toPromise();
+        })
+        .then((entities) => {
+          expect(entities).toEqual([]);
+          return store.pendingSyncCount();
+        })
+        .then((count) => {
+          expect(count).toEqual(2);
         });
     });
   });
 
   describe('pendingSyncCount()', () => {
     it('should return the count of entities waiting to be synced', () => {
-      const store = new CacheStore(collection);
-      const syncStore = new SyncStore(collection);
+      const store = new SyncStore(collection);
       const entity = {};
 
-      return syncStore.save(entity)
+      return store.save(entity)
         .then((entity) => {
           const query = new Query().equalTo('_id', entity._id);
           return store.pendingSyncCount(query);
@@ -1033,11 +668,10 @@ describe('CacheStore', () => {
 
   describe('pendingSyncEntities()', () => {
     it('should return the entities waiting to be synced', () => {
-      const store = new CacheStore(collection);
-      const syncStore = new SyncStore(collection);
+      const store = new SyncStore(collection);
       const entity = {};
 
-      return syncStore.save(entity)
+      return store.save(entity)
         .then((entity) => {
           const query = new Query().equalTo('_id', entity._id);
           return store.pendingSyncEntities(query)
@@ -1051,16 +685,15 @@ describe('CacheStore', () => {
     });
   });
 
-  describe('push()', () => {
+  describe('push', () => {
     it('should push the entities to the backend', () => {
-      const store = new CacheStore(collection);
-      const syncStore = new SyncStore(collection);
+      const store = new SyncStore(collection);
       const entity = { _id: randomString() };
 
-      return syncStore.save(entity)
+      return store.save(entity)
         .then(() => {
-          nock(store.client.apiHostname)
-            .put(`${store.pathname}/${entity._id}`, entity)
+          nock(client.apiHostname)
+            .put(`/appdata/${client.appKey}/${collection}/${entity._id}`, entity)
             .reply(200, entity);
 
           return store.push();
@@ -1075,16 +708,17 @@ describe('CacheStore', () => {
     });
 
     it('should push only the entities matching the query to the backend', () => {
-      const store = new CacheStore(collection);
-      const syncStore = new SyncStore(collection);
+      const store = new SyncStore(collection);
       const entity1 = { _id: randomString() };
       const entity2 = { _id: randomString() };
 
-      return syncStore.save(entity1)
-        .then(() => syncStore.save(entity2))
+      return store.save(entity1)
         .then(() => {
-          nock(store.client.apiHostname)
-            .put(`${store.pathname}/${entity1._id}`, entity1)
+          return store.save(entity2);
+        })
+        .then(() => {
+          nock(client.apiHostname)
+            .put(`/appdata/${client.appKey}/${collection}/${entity1._id}`, entity1)
             .reply(200, entity1);
 
           const query = new Query().equalTo('_id', entity1._id);
@@ -1100,11 +734,11 @@ describe('CacheStore', () => {
     });
   });
 
-  describe('pull()', () => {
+  describe('pull', () => {
     it('should save entities from the backend to the cache', () => {
       const entity1 = { _id: randomString() };
       const entity2 = { _id: randomString() };
-      const store = new CacheStore(collection);
+      const store = new SyncStore(collection);
 
       nock(store.client.apiHostname)
         .get(`/appdata/${store.client.appKey}/${collection}`)
@@ -1121,21 +755,20 @@ describe('CacheStore', () => {
     });
   });
 
-  describe('sync()', () => {
+  describe('sync', () => {
     it('should push any pending sync entities and then pull entities from the backend and save them to the cache', () => {
-      const store = new CacheStore(collection);
-      const syncStore = new SyncStore(collection);
+      const store = new SyncStore(collection);
       const entity1 = { _id: randomString() };
       const entity2 = { _id: randomString() };
 
-      return syncStore.save(entity1)
+      return store.save(entity1)
         .then(() => {
-          nock(store.client.apiHostname)
-            .put(`${store.pathname}/${entity1._id}`, entity1)
+          nock(client.apiHostname)
+            .put(`/appdata/${client.appKey}/${collection}/${entity1._id}`, entity1)
             .reply(200, entity1);
 
           nock(store.client.apiHostname)
-            .get(store.pathname)
+            .get(`/appdata/${store.client.appKey}/${collection}`)
             .reply(200, [entity1, entity2]);
 
           return store.sync();
@@ -1153,11 +786,10 @@ describe('CacheStore', () => {
 
   describe('clearSync()', () => {
     it('should clear the sync table', () => {
-      const store = new CacheStore(collection);
-      const syncStore = new SyncStore(collection);
+      const store = new SyncStore(collection);
       const entity = { _id: randomString() };
 
-      return syncStore.save(entity)
+      return store.save(entity)
         .then(() => {
           return store.clearSync();
         })
@@ -1166,6 +798,7 @@ describe('CacheStore', () => {
         })
         .then((count) => {
           expect(count).toEqual(0);
+          const syncStore = new SyncStore(collection);
           const query = new Query().equalTo('_id', entity._id);
           return syncStore.find(query).toPromise();
         })
@@ -1175,13 +808,14 @@ describe('CacheStore', () => {
     });
 
     it('should clear only the entities from the sync table matching the query', () => {
-      const store = new CacheStore(collection);
-      const syncStore = new SyncStore(collection);
+      const store = new SyncStore(collection);
       const entity1 = { _id: randomString() };
       const entity2 = { _id: randomString() };
 
-      return syncStore.save(entity1)
-        .then(() => syncStore.save(entity2))
+      return store.save(entity1)
+        .then(() => {
+          return store.save(entity2);
+        })
         .then(() => {
           const query = new Query().equalTo('_id', entity1._id);
           return store.clearSync(query);
@@ -1192,6 +826,7 @@ describe('CacheStore', () => {
         })
         .then((count) => {
           expect(count).toEqual(0);
+          const syncStore = new SyncStore(collection);
           const query = new Query().equalTo('_id', entity1._id);
           return syncStore.find(query).toPromise();
         })
